Refactor demo to home
