+++ frontend/src/pages/home/aws_association.json
