+++ frontend/src/pages/home/azure_association.json
