+++ frontend/src/pages/home/azure_demo_new.json
