+++ frontend/src/pages/index.js
@@ -1,4 +1,4 @@
-export { default as Demo } from './demo/demo';
-export { default as About } from './about/about';
-export { default as Contact } from './contact/contact';
-export { default as Document } from './document/document';
+export { default as Home } from "./home/home";
+export { default as About } from "./about/about";
+export { default as Contact } from "./contact/contact";
+export { default as Document } from "./document/document";
